airflow-docker/requirements.txt
boto3
scrapy
pymysql
sqlalchemy
splash
scrapy-splash
Pillow
html-to-json
scrapy-deltafetch
pyspark
wheel
pandas
pyarrow
pandas
numpy
cyrtranslit
tabula-py
apache-airflow==2.4.3
findspark
sqlacodegen
pydantic
mysql-connector-python
xlrd
git-remote-codecommit
apache-airflow-providers-amazon
apache-airflow-providers-mysql
requests
apache-airflow[amazon]
mysql-replication
pydantic[dotenv]
python-dotenv
scikit-learn
lightgbm
xgboost
statsmodels
scikit-plot
apache-airflow-providers-apache-spark
